src/web/requirements.txt
streamlit==1.35.0
langchain==0.2.5
azure-identity==1.16.1
langchain-openai==0.1.8
python-dotenv==1.0.1
langchainhub==0.1.20
pandas==2.2.2
openai==1.34.0
langchain-community==0.2.5
pandas==2.2.2
pytz==2024.1